Extract tests related to dashboard filters with nested questions to a separate file
